Removed a extra file of language selection screen.
